third_party/WebKit/LayoutTests/animations/prefixed/animation-unprefixed-string.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  @keyframes anim {
    from { background: red; }
    to { background: red; }
  }

  #target {
    animation: "anim" 1s forwards;
    background: green;
  }
</style>
<div id="target"></div>
<script>
test(function() {
  assert_equals(getComputedStyle(target).backgroundColor, 'rgb(0, 128, 0)');
}, 'Check that unprefixed quoted animation name is illegal');
</script>
